# lib/api-zod/src/generated/types/cutoutRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CutoutRequest {
  /** Base64-encoded source image (no data URI prefix) */
  image: string;
  /** MIME type of the source image (e.g. image/jpeg) */
  mimeType: string;
}