# lib/api-zod/src/generated/types/cutoutResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CutoutResult {
  /** Base64-encoded PNG cut-out with transparent background */
  image: string;
  /** MIME type of the returned image */
  mimeType: string;
}